Added missing cast.

+++ Magick.NET/Core/Types/BigRational.cs
@@ -147,7 +147,7 @@ namespace ImageMagick
       Denominator = 1;
 
       double val = Math.Abs(value);
-      double df = Numerator / Denominator;
+      double df = Numerator / (double)Denominator;
       double epsilon = bestPrecision ? double.Epsilon : .000001;
 
       while (Math.Abs(df - val) > epsilon)
